refactor: auth redirect logic as auth guard
